+++ roles/jd-server/Cargo.toml
@@ -5,6 +5,10 @@ edition = "2018"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/stratum-mining/stratum"
 
+[lib]
+name = "jd_server"
+path = "src/lib/mod.rs"
+
 [dependencies]
 stratum-common = { path = "../../common" }
 async-channel = "1.5.1"
