+++ roles/jd-server/src/lib/job_declarator/message_handler.rs
@@ -17,7 +17,7 @@ use roles_logic_sv2::errors::Error;
 use stratum_common::bitcoin::consensus::Decodable;
 use tracing::warn;
 
-use crate::lib::job_declarator::signed_token;
+use super::signed_token;
 use stratum_common::bitcoin::consensus::encode::serialize;
 
 use super::JobDeclaratorDownstream;
