+++ roles/jd-server/src/lib/status.rs
@@ -1,6 +1,6 @@
 use roles_logic_sv2::parsers::Mining;
 
-use crate::error::JdsError;
+use super::error::JdsError;
 
 /// Each sending side of the status channel
 /// should be wrapped with this enum to allow
